+++ .babelrc
@@ -1,0 +1,8 @@
+{
+    //"presets": ["es2015", "react"],
+    "env": {
+        "test": {
+          "plugins": ["@babel/plugin-transform-modules-commonjs"]
+        }
+      }
+  }
