+++ package.json
@@ -20,9 +20,27 @@
 		"dotenv": "^14.3.2",
 		"express": "^4.17.2",
 		"md5": "^2.3.0",
-		"pg": "^8.7.1",
 		"nodemailer": "^6.7.2",
-		"jest": "^27.5.1"
+		"pg": "^8.7.1"
+	},
+	"devDependencies": {
+		"@babel/plugin-transform-modules-commonjs": "^7.17.9",
+		"@types/jest": "^27.4.1",
+		"babel-jest": "^27.5.1",
+		"jest": "^27.5.1",
+		"supertest": "^6.2.2"
+	},
+	"jest": {
+		"verbose": true,
+		"forceExit": true,
+		"testPathIgnorePatterns": [
+			"/node_modules/"
+		],
+		"moduleFileExtensions": [
+			"js",
+			"json",
+			"es6"
+		]
 	},
 	"engines": {
 		"node": "16.x"
